Add files via upload
This is the app including a start up page. I was intending to have helper classes (i.e., RestaurantMapper) handle the API queries but I was having problem with the context portion. I still think this can be done so I left the helper class in for future refactoring (time permitting). Also, getting the device location from the GPS has a couple of steps. I didn't get into it because I wanted to be able to get the API and ListView updates running. I will look into that again but I left a source in the comments describing how to do it.